Started working on the coding part
We made some changes to the the htmls, mainly name changes or adding data-attribute where it would be necessary, created file called initial page load to make all html elements, created classes.js to store todo, project, and note class that we made using composition. However, we aren't going to use the 'Note' class, it's just for demo right now. Imported functions to index.js and created a module which calls imported functions to store all important html elements in a module for later use. Next commits  will be working on the functionality for loading the modal, and the functionality for visually indicating what tabs the user is on.