Major changes
Changed element ids for modal content in createModalContent to be more specific, added functions to see if dates were the same or within 7 days of each other since soon we'll be making it so the main tabs show todos due today or within 7 days, however this hasn't been tested out yet. Changed element ids in initialPageLoad to be more specific rather than having classes. Also removed content hidden from the functions that created the overlay and modal, 'content-hidden' would be added in index.js since it felt unusual just already having them hidden. Finished function for creating the main content section. New data attributes called tabtype help differentiate between a main tab 'home' 'today' or 'week' from a tab for a project. The tabID data attribute is only for main tabs and they indicate whether it's the home, todya, or week tab. For buttons that create/edit projects or todos, we gave them a data-attribute called for-action, which helps us differentiate whether the user is wanting to creating a new instance or edit an existing one. In index.js we started just creating the logic for displaying a project form, the logic for submitting the form, code for updating the sidebar tabs, and logic for rendering mai ncontent