Finished base functionality for project
Finished functionality for displaying todo form, for showing todo details. Finished rendering different todo elements for a project tab and a main tab. We finished functionality for toggling the completion of a todo, editing, and deleting the todo. Also gave functionality to deleting a project, and also looks like we finished the small functionfor editing a project. In next commit we will be mainly working on cleaning up and changing code to make it render. After that we'll work on adding local storage.