Finished functionality
Created one boolean or value called 'isEdit' to keep track of whether a user is editing or adding something, and removed the previous projectFormState and todoFormState. Created functions for showing dates in U.S. format time and one for sorting todos by ascending date. Since updateSidebarTabs and renderMainContent are always called together, we created a function for calling those two functions called 'renderPage'. Removed data-attribute data-form-action, since we have edit boolean to keep track of things now. Finished functionality for showing and ordering todos chronologically. Finished functionality for only including incomplete todos in the count on the sidebar. Added functionality so that if there are no todos in a certain tab, we will show a message telling hte user to 'add a todo or create a project to get started'. Finished functionality to show due date on todos, and finished/polished functionality to show when a todo is finished or not, with its name being crossed out or the 'toggleCompletionBtn' having different text, done in renderMainContent. Add event listeners for adding todo to project, editing project, deleting projects, toggling todo completion, editing, and deleting todos. Finally, we added localStorage so that the application saves user changes when the user adds/edits/deletes projects or todos everytime they do it. Next we will be working on cleaning up the css and adding a dark theme button