Finishing touches
Did some finishing touches and deleted unnecessary files. Decided not to do the dark-mode but I really want to try to do the authentication thing in the future and have two separate parts for the application, one part for loading in a demo and another part for loading in some user's todos. Anyways that was satisfying.