Started preparing for rendering and styling.
Created initialPage.js to create all of the markup. We've created modules.js to contain all of our data data-structures and modules that we'll use throughout the application. There were some changes to the style.css for the forms and modals. copyIndex.js will be a file we'll reference if we need help remembering how the application should work. So next we'll be working on rendering and hopefully adding event listeners to our page.