Main content rendering going well
We started creating our rendering logic in rendering.js. Finished rendering logic for modal logic and how to display the individual forms or sections in the modal. Also finished logic for sidebar and how to render main content. Started working on pageListeners.js, which is where our we create the callback functions for our event listeners, which we'll import to rendering.js, if needed. Then in modules.js we got references to various dom elements or functions that query for certain dom elements when dom elements are added dynamically. More importantly we have tabsModule, which is an important object that contains infomration about which tabs are active and so forth. With our loadDummyData function, we've successfully tested that our rendering logic for todos and projects is working. Next we'll work on more event listeners