Improvement to webpack setup and other fixes
Improved the webpack setup, now it uses a bundle.js and whatnot. Removed unnecsesary assets, and imported images. Finally I'm probably going to test it and make sure nothing's wrong, after I'm going to write up some small documentation on the architecture of the application and how the application should be handled if we need to change it